Got training and classification working decently.  Time to run experiments.

+++ src/Detector.h
@@ -17,7 +17,7 @@ public:
             bool isPositive,
             bool shouldShowHogImage = false );
 
-    float classify(
+    double classify(
             const cv::Mat& image,
             bool shouldShowHogImage = false );
